[refactor] Refactors MyStory component css into separate components.
