+++ pages/index.tsx
@@ -10,7 +10,8 @@ import { ProjectsContainer } from '../src/components/Projects/ProjectsContainer'
 import { projectsList } from '../src/content/projects';
 import { ProjectFull } from '../src/components/Projects';
 import { ProjectsHeader } from '../src/components/Projects/ProjectsHeader';
-import { MyStory, WorkExperience } from '../src/components/About';
+import { WorkExperience } from '../src/components/WorkExperience/WorkExperience';
+import { MyStory } from '../src/components/MyStory/MyStory';
 import { DetailBlock } from '../src/components/DetailBlock';
 import { DetailsBlockContainer } from '../src/components/Projects/DetailsBlockContainer';
