--- src/components/About/index.tsx
@@ -1,4 +1,0 @@
-import { MyStory } from './MyStory';
-import { WorkExperience } from './WorkExperience';
-
-export { MyStory, WorkExperience };
